+++ src/collective/volto/dropdownmenu/restapi/dropdown_menu/get.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 from collective.volto.dropdownmenu.interfaces import IDropDownMenu
+from collective.volto.dropdownmenu.restapi.serializer.dropdown_menu import (
+    serialize_data,
+)
 from plone import api
 from plone.restapi.services import Service
 from zope.interface import implementer
 from zope.publisher.interfaces import IPublishTraverse
-
-import json
 
 
 @implementer(IPublishTraverse)
@@ -18,5 +19,5 @@ class DropDownMenuGet(Service):
             "menu_configuration", interface=IDropDownMenu, default=""
         )
         if not record:
-            return {}
-        return json.loads(record)
+            return []
+        return serialize_data(json_data=record, show_children=True)