+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Customized controlpanel's serializer and deserializer to handle internal links with uids.
+  [cekk]
 
 
 0.1.0 (2020-06-11)